Changes to error logging for forking users
Previous setup required the bot to be in the original developer's bot server for logging errors.

A potential better solution could be to put the guild_id of the testing server as well as the channel_id of the error channel and online channel in the .env file, and read them from there to allow forkers to use the error logging too.